+++ stories/components/data-table/docs/examples/accordion-inside-table.tsx
@@ -29,6 +29,18 @@ const Demo = (props: AccordionInTableProps) => {
           ],
         },
       ]}
+      renderCell={(props) => {
+        const parentRowIndex = props.rowIndex - 1;
+
+        if (parentRowIndex === 0 && props.columnName === ACCORDION) {
+          return {
+            p: 0, // set empty paddings for the first accordion
+            children: props.defaultRender(),
+          };
+        }
+
+        return props.defaultRender();
+      }}
     />
   );
 };