+++ semcore/data-table/CHANGELOG.md
@@ -21,6 +21,7 @@ CHANGELOG.md standards are inspired by [keepachangelog.com](https://keepachangel
 - Accordion content expands with visual issues when clicking on a merged cell.
 - Incorrect type for `selectedRows/onSelectedRowsChange` when `uniqueRowKey` type is not a string.
 - Hover styles aren't applied to themed cells.
+- Types for `CellRenderProps` - added ability to compare props.columnName with `ACCORDION` symbol.
 
 ### Changed